style: refactor selector options
