+++ packages/glyph-components/src/components/selector/readme.md
@@ -33,12 +33,13 @@
 
 ### Depends on
 
-- [glyph-input](../input)
+- [glyph-selector-options](options)
 
 ### Graph
 ```mermaid
 graph TD;
-  glyph-selector --> glyph-input
+  glyph-selector --> glyph-selector-options
+  glyph-selector-options --> glyph-input
   glyph-timeline --> glyph-selector
   glyph-user-menu --> glyph-selector
   style glyph-selector fill:#f9f,stroke:#333,stroke-width:4px
