+++ packages/glyph-components/src/components/selector/selector.stories.mdx
@@ -55,7 +55,7 @@ options = {
 
 
 <Canvas>
-    <Story name="Comples" args={{ 'multi-select': true, label: 'simple', '.complexOptions': { group1: [{ name: 'option 1', active: false, value: 1 }, { name: 'option 2', active: false, value: 2 }],
+    <Story name="Complex" args={{ 'multi-select': true, label: 'simple', '.complexOptions': { group1: [{ name: 'option 1', active: false, value: 1 }, { name: 'option 2', active: false, value: 2 }],
 group2: [{ name: 'option 3', active: false, value: 3 }, { name: 'option 4', active: false, value: 4 }]} }}>
         {ComplexTemplate.bind({})}
     </Story>
@@ -77,6 +77,7 @@ Selector can be used with a search input to filter options easily.
 <Description>{notes}</Description>
 
 export const changeEvent = (event) => {
+console.log(event);
     event.target.options = event.target.options.map(option => {
         if (option.value === event.detail.value) {
             if (event.target.multiSelect) {
@@ -91,6 +92,7 @@ export const changeEvent = (event) => {
     })
 }
 export const changeEventComplex = (event) => {
+    console.log(event);
     event.target.complexOptions = Object.keys(event.target.complexOptions).reduce((prev, curr) => {
         prev[curr] = event.target.complexOptions[curr].map(option => {
             if (option.value === event.detail.value) {
