+++ packages/glyph-components/src/components/selector/selector.tsx
@@ -1,4 +1,4 @@
-import { Component, Prop, h, State, Event, EventEmitter, Listen, Element } from '@stencil/core';
+import { Component, Prop, h, State, Event, EventEmitter, Element } from '@stencil/core';
 import { ComplexSelectorOptions, SelectorOption, UIInterface } from 'glyph-core';
 import { Icon } from '../../utils/icons';
 import { Flex } from '../../utils/layout';
@@ -25,24 +25,13 @@ export class SelectorComponent {
     /** on change callback */
     @Event() optionSelect: EventEmitter<SelectorOption>;
 
-    /** Element reference */ 
+    /** Element reference */
     @Element() element: HTMLGlyphSelectorElement;
 
     /** Selected option */
     @State() selectedOptions: SelectorOption[];
     /** Options selector drilldown flag */
     @State() optionsDrilldown: boolean;
-    /** Search text */
-    @State() searchValue: string;
-
-    @Listen('click', { target: 'window' })
-    clickOutside(event: any) {
-        if (this.optionsDrilldown) {
-            if (!event.composedPath().includes(this.element)) {
-                this._toggleContainer();
-            }
-        }
-    }
 
     componentWillRender() {
         if (this.options) {
@@ -55,113 +44,18 @@ export class SelectorComponent {
         }
     }
 
-    private _toggleContainer = () => {
-        this.optionsDrilldown = !this.optionsDrilldown;
-
-        if (!this.optionsDrilldown) {
-            this.searchValue = undefined;
-        }
+    private _toggleContainer = (value?: boolean) => () => {
+        this.optionsDrilldown = value ?? !this.optionsDrilldown;
     };
 
-    private _selectOption = (option: SelectorOption) => () => {
+    private _selectOption = (option: SelectorOption) => {
         this.optionSelect.emit(option);
-        if (!this.multiSelect) {
-            this.optionsDrilldown = false;
-        }
-    };
-
-    private _handleInputChange = (event: any) => {
-        this.searchValue = event.detail;
-    };
-
-    private _handleEnter = () => {
-        let visibleOptions = [];
-        if (this.complexOptions) {
-            visibleOptions = Object.keys(this.complexOptions).reduce(
-                (acc, key) => acc.concat(this.complexOptions[key].filter(this._inSearch)),
-                [],
-            );
-        }
-        if (this.options) {
-            visibleOptions = this.options.filter(this._inSearch);
-        }
-        if (visibleOptions.length === 1) {
-            this._selectOption(visibleOptions[0])();
-        }
-    };
-
-    private _renderSearch = () => {
-        return (
-            <glyph-input
-                autoFocus
-                box
-                placeholder={this.searchPlaceholder}
-                onEnterKey={this._handleEnter}
-                onTextChange={this._handleInputChange}
-            />
-        );
-    };
-
-    private _renderCheckbox = (option: SelectorOption) => {
-        return (
-            <Flex middle center className={cls('checkbox', option.active && 'active')}>
-                {option.active && <Icon icon="checkmark" />}
-            </Flex>
-        );
-    };
-
-    private _inSearch = (option: SelectorOption) => {
-        if (this.searchValue) {
-            return option.name.toLowerCase().includes(this.searchValue.toLowerCase());
-        }
-
-        return true;
-    };
-
-    private _renderComplexOptions = () => {
-        return (
-            <Flex row>
-                {Object.keys(this.complexOptions).map(key => (
-                    <Flex>
-                        <label>{key}</label>
-                        {this._renderOptions(this.complexOptions[key])}
-                    </Flex>
-                ))}
-            </Flex>
-        );
-    };
-
-    private _renderSimpleOptions = () => {
-        return <Flex>{this._renderOptions(this.options)}</Flex>;
-    };
-
-    private _renderOptions = (options: SelectorOption[] = this.options) => {
-        return (
-            <Flex>
-                {options
-                    .filter(option => this._inSearch(option))
-                    .map((option: SelectorOption) => (
-                        <Flex className="option" row middle left onClick={this._selectOption(option)}>
-                            {this.multiSelect && this._renderCheckbox(option)}
-                            {this._renderOptionName(option)}
-                        </Flex>
-                    ))}
-            </Flex>
-        );
-    };
-
-    private _renderOptionName = ({ name }: SelectorOption) => {
-        if (this.searchValue) {
-            name = name.split(this.searchValue).join(`<b>${this.searchValue}</b>`);
-        }
-
-        return <span innerHTML={name} />;
     };
 
     render() {
         return (
             <Flex className={cls('selector__container', this.interface)}>
-                <Flex row spaced className="selector__input" onClick={this._toggleContainer}>
+                <Flex row spaced className="selector__input" onClick={this._toggleContainer()}>
                     <label class={cls(this.selectedOptions?.length && 'active')}>{this.label}</label>
                     <span>{this.selectedOptions?.map(option => option.name).join(', ') || ''}</span>
                     <Icon
@@ -169,11 +63,16 @@ export class SelectorComponent {
                         className={cls(this.optionsDrilldown && 'active')}
                     />
                 </Flex>
-                <Flex className="selector__options__container">
-                    {this.optionsDrilldown && this.searchPlaceholder && this._renderSearch()}
-                    {this.optionsDrilldown && this.options && this._renderSimpleOptions()}
-                    {this.optionsDrilldown && this.complexOptions && this._renderComplexOptions()}
-                </Flex>
+                {this.optionsDrilldown && (
+                    <glyph-selector-options
+                        options={this.options}
+                        complexOptions={this.complexOptions}
+                        multiSelect={this.multiSelect}
+                        searchPlaceholder={this.searchPlaceholder}
+                        optionClickEvent={this._selectOption}
+                        closeEvent={this._toggleContainer(false)}
+                    />
+                )}
             </Flex>
         );
     }
